Revise MVC4Support Area Controller Selector
Due the original implementation will search the full loaded assemblies
to find the controller, there is NHibernate.Linq that throw exception
when do tranverse and reflection. So I have to change the search logic
to use Spring.Context to find the injected controllers.

Add the OData implementation demo for RolesController